Unify all OrdersController in one single file
