+++ csc-services/controllers/impl/index.js
@@ -4,3 +4,4 @@ require("./TableTemplateControllerImpl");
 require("./HeaderControllerImpl");
 require("./NotificationsControllerImpl");
 require("./DashboardMenuControllerImpl");
+require("./OrdersControllerImpl");
