+++ csc-site-ssapp/code/scripts/app-config.js
@@ -9,6 +9,7 @@ const HistoryModalController = cscServices.getController('HistoryModalController
 const HeaderController = cscServices.getController('HeaderController', ACTOR);
 const NotificationsController = cscServices.getController('NotificationsController', ACTOR);
 const DashboardMenuController = cscServices.getController('DashboardMenuController', ACTOR);
+const OrdersController = cscServices.getController('OrdersController', ACTOR);
 
 addControllers({
 	TableTemplateController,
@@ -16,5 +17,6 @@ addControllers({
 	HistoryModalController,
 	HeaderController,
 	NotificationsController,
-	DashboardMenuController
+	DashboardMenuController,
+	OrdersController
 });
